refactor(silencer): split in-game lifecycle into ingame.cpp
Phase 4 mechanical move. LoadMap, UnloadGame, JoinGame,
GiveDefaultItems, ShowDeployMessage, CheckForQuit/EndOfGame/
ConnectionLost, ProcessInGameInterfaces, and ShowTeamOverlays now
live in src/game/ingame.cpp. Declarations stay on Game in game.h.
No behavior change.
